Blogit toimii nyt reduxilla, loggedUseria aloiteltu

+++ bloglist-frontend/src/components/Blog.js
@@ -8,7 +8,6 @@ import { deleteBlog } from '../reducers/blogReducer'
 class Blog extends React.Component {
   constructor(props) {
     super(props)
-    console.log('Nyt ollaan Blog konstruktorissa')
     this.state = {
       showAll: false
     }
@@ -22,7 +21,6 @@ class Blog extends React.Component {
 
   handleLike = async (blog) => {
     await this.props.addLike(blog)
-    console.log('Blogi handleLikessa: ' + blog.id)
     this.props.addSuccessNotification('tänne tekstiä')
     setTimeout(() => {
       this.props.addSuccessNotification(null)
@@ -30,27 +28,28 @@ class Blog extends React.Component {
   }
 
   handleDelete = async (blog) => {
-    const result = this.props.deleteBlog(blog)
-    if (result === "error") {
-      this.props.addErrorNotification('Can\'t delete other user\'s blogs')
-      setTimeout(() => {
-        this.props.addErrorNotification(null)
-      }, 5000)
-    } else if (result !== null) {
-      this.props.addSuccessNotification(`${blog.title} poistettu!`)
-      setTimeout(() => {
-        this.props.addSuccessNotification(null)
-      }, 5000)
+    if (window.confirm('Poistetaanko \'' + blog.title + '\' by ' + blog.author + '?')) {
+      const response = await this.props.deleteBlog(blog)
+      console.log(response)
+      if (response !== "error") {
+        this.props.addSuccessNotification(`${blog.title} poistettu!`)
+        setTimeout(() => {
+          this.props.addSuccessNotification(null)
+        }, 5000)
+      } else {
+        this.props.addErrorNotification(`Can't remove other user's blogs!`)
+        setTimeout(() => {
+          this.props.addErrorNotification(null)
+        }, 5000)
+      }
     }
   }
 
   render() {
     console.log('renderöidään Blogissa')
-    console.log('renderöitävän blogin id: ' + this.props.blog.id)
-    /*Noihin alempiin voidaan laittaa toisena ehtona,
-    jos this.props.blog.id === this.props.recentlyLiked, niin display: ''*/
     const showAllInfo = { display: this.state.showAll ? '' : 'none' }
     const onlyShowTitleAndAuthor = { display: this.state.showAll ? 'none' : '' }
+    /*Jos nyt sattuis käymään niin että delete näkyis muillekki, nii ei ne niitä poistaa voi ja saavat ilmotuksen*/
     const showDelete = { display: (this.props.user.id === this.props.blog.user._id) ? '' : 'none'}
     const blogStyle = {
       paddingTop: 10,
@@ -58,18 +57,6 @@ class Blog extends React.Component {
       border: 'solid',
       borderWidth: 1,
       marginBottom: 5
-    }
-    const blog = {
-      id: this.props.blog.id,
-      title: this.props.blog.title,
-      author: this.props.blog.author,
-      url: this.props.blog.url,
-      likes: this.props.blog.likes,
-      user: {
-        _id: this.props.blog.user._id,
-        username: this.props.blog.user.username,
-        name: this.props.blog.user.name
-      }
     }
     return (
       <div style={blogStyle}>
